+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 Fixed
 =====
 
+- Prevent flickering in ``euporie-hub`` clients on server log output
 - Notify of dead kernel immediately
 - Handle non-existent file in the text file editor
 - Fix issue with Select widget not scrolling beyond selection with scrollbar